v0.2

+++ setup.py
@@ -6,6 +6,32 @@ import re
 import sys
 import os
 
+__ascii_art__ = """
+                    ....
+              ....--::::-==-:.        .::..
+           ..---::.         :=:..    .:  :..
+        ...--.     ..... ..   :-::   .:   .:.
+       :::-.   .---:::::-::---  :=:.   :    :.
+     .-=.   .---..          .:+   =.. ..:    :.
+   .:=:    --:.               :=  .=..  .:    :.
+  :-=    -=:.                 .:=  + :  .:    :.
+ :.=    --.   .:..:.           :+ --:    .:   :.
+..+. .--:.  .:.   .:  .        :.-:.     .:   ...
+ .:--:..  .::   .:.            ...       .:   ...
+         . :  .:.                        .:   : .
+        ..: .. .                        .:    :.
+        ... ...                        .:    :..
+        .:   :.         .             .:   :
+        .:    :.        .....:.. ..  ..   ...
+         .:    :..      ...    .. .........
+          .:    :.        ............
+           .:.    .:.
+             .::     ..::.   ...
+               .:..       .... ..:.
+                  ......          .-.
+                        .:::.....::..
+"""
+
 # set logger
 logging.basicConfig(level=logging.INFO, format='%(message)s')
 
@@ -13,29 +39,32 @@ logging.basicConfig(level=logging.INFO, format='%(message)s')
 py_version = sys.version_info
 if py_version < (3, 7):
     sys.exit("Python < 3.7 is not supported, aborting setup")
-logging.info("Running setup with Python {}.{}".format(py_version.major, py_version.minor))
 
 # get directory path
 dir_path = os.path.dirname(os.path.realpath(__file__))
 
 # get version from __init__
-init_string = open(os.path.join(dir_path, 'bajes', '__init__.py')).read()
-VERS = r"^__version__ = ['\"]([^'\"]*)['\"]"
-mo = re.search(VERS, init_string, re.M)
+inis    = open(os.path.join(dir_path, 'bajes', '__init__.py')).read()
+VERS    = r"^__version__ = ['\"]([^'\"]*)['\"]"
+mo      = re.search(VERS, inis, re.M)
 VERSION = mo.group(1)
-                     
-# Tidy up the project root
-CLEAN_FILES = './build ./dist ./*.pyc ./*.tgz ./*.egg-info'.split(' ')
 
-for path_spec in CLEAN_FILES:
-    # Make paths absolute and relative to this path
-    abs_paths = glob.glob(os.path.normpath(os.path.join(dir_path, path_spec)))
-    for path in [str(p) for p in abs_paths]:
-        if not path.startswith(dir_path):
-            # Die if path in CLEAN_FILES is absolute + outside this directory
-            raise ValueError("{} is not a path inside {}".format(path, dir_path))
-        logging.info("removing {}".format(os.path.relpath(path)))
-        shutil.rmtree(path)
+# Tidy up the project root if installing
+if 'install' in sys.argv:
+
+      logging.info("Running Bajes {} installation with Python {}.{}".format(VERSION, py_version.major, py_version.minor))
+      logging.info("cleaning project root")
+      CLEAN_FILES = './build ./dist ./*.pyc ./*.tgz ./*.egg-info'.split(' ')
+
+      for path_spec in CLEAN_FILES:
+          # Make paths absolute and relative to this path
+          abs_paths = glob.glob(os.path.normpath(os.path.join(dir_path, path_spec)))
+          for path in [str(p) for p in abs_paths]:
+              if not path.startswith(dir_path):
+                  # Die if path in CLEAN_FILES is absolute + outside this directory
+                  raise ValueError("{} is not a path inside {}".format(path, dir_path))
+              logging.info("removing {}".format(os.path.relpath(path)))
+              shutil.rmtree(path)
 
 setup(# metadata
       name='bajes',
@@ -43,14 +72,20 @@ setup(# metadata
       description='Bayesian Jenaer Software',
       long_description=open(os.path.join(dir_path, 'README.md')).read(),
       long_description_content_type="text/x-rst",
-      author='Matteo Breschi, Rossella Gambda, Sebastiano Bernuzzi',
+      author='Matteo Breschi, Rossella Gamba, Sebastiano Bernuzzi et al.',
       author_email='matteo.breschi@uni-jena.de',
       url='https://github.com/matteobreschi/bajes',
       license='MIT',
-      
-      # list of packages and dirs
-      packages  = find_packages(),
-      
+
+      # list of packages and data
+      packages=find_packages(),
+      package_dir={"bajes": "bajes"},
+      package_data={"bajes": ["pipe/data/gw/asd/events/*/*.txt",
+                              "pipe/data/gw/asd/design/*.txt",
+                              "pipe/data/gw/spcal/events/*/*.txt",
+                              "pipe/data/kn/filter/AT2017gfo/*.txt",
+                              "obs/kn/fluxfactors/*.dat"]},
+
       # make scripts executable
       scripts   = ['bajes/pipe/scripts/bajes_core.py',
                    'bajes/pipe/scripts/bajes_parallel_core.py',
@@ -58,18 +93,14 @@ setup(# metadata
                    'bajes/pipe/scripts/bajes_inject.py',
                    'bajes/pipe/scripts/bajes_read_gwosc.py',
                    'bajes/pipe/scripts/bajes_postproc.py'],
-      
-      # include data
-      package_data={"bajes": ["pipe/data/gw/asd/events/*/*.txt",
-                              "pipe/data/gw/asd/design/*.txt",
-                              "pipe/data/gw/spcal/events/*/*.txt",
-                              "pipe/data/kn/filter/AT2017gfo/*.txt",
-                              "obs/kn/fluxfactors/*.dat"]},
-      
+
       # set mandatory requirements
       python_requires='>=3.7',
       install_requires=['numpy>=1.18.0',
                         'scipy>=1.4.0',
                         'astropy>=4.0.0']
       )
-logging.info("Bajes succesfully installed.")
+
+if 'install' in sys.argv:
+      logging.info("Bajes succesfully installed.")
+      logging.info(__ascii_art__)
